Added special treatment for catalog_category_view regarding parameters
To make the module work better out of the box for layered category pages,
I added an option to include all parameters in the cache key of a category
view page.

Also, you can configure what the maximum number of parameters is for a
page to be saved. This way the cache size cannot grow exponentially with
a large number of filterable attributes and still have the frequently
visited pages cached.